# src/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "This ios Intellidrone API designed by [Singleton](http://singleton.com.ar).You\
    \  can find out more about Swagger at [Singleton Web Page](http://singleton.com.ar)\
    \ or on [Intellidrone Web Page](http://robotagro.com)."
  version: "1.0.0"
  title: "Intellidrone API"
  termsOfService: "http://web.robotagro.com/api//terms"
  contact:
    name: "Equipo de Soporte"
    url: "http://singleton.com.ar/soporte"
    email: "soporte@singleton.com.ar"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:8080"
basePath: "/"
...
schemes:
- "http"
- "https"
paths:
...
  /Actividades:
    get:
      tags:
      - "Actividades"
      summary: "Get all Actividades."
      description: "Get all Actividades."
      operationId: "getActividades"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "skip"
        in: "query"
        description: "number of item to skip (not included the first)"
        required: true
        type: "integer"
        format: "int32"
      - name: "limit"
        in: "query"
        description: "max records to return (included the last)"
        required: true
        type: "integer"
        format: "int32"
      - name: "orderBy"
        in: "query"
        description: "order by property. Put '-' at the beginning to descendant order."
        required: false
        type: "string"
      - name: "filter"
        in: "query"
        description: "filter data. NOT SUPPORTED YET."
        required: false
        type: "string"
      - name: "userId"
        in: "query"
        description: "id of user. Only for admin users."
        required: false
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Request OK."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Vacas"
        204:
          description: "Response is empty."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
        400:
          description: "Some parameters are missing or badly entered."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "string"
              format: "date-time"
              description: "The UTC date/time at which the current rate limit window\
                \  resets."
          schema:
            type: "string"
        404:
          description: "Entity not found."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Illegal input for operation."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
      security:
      - appToken: []
      - userToken: []
      x-swagger-router-controller: "Actividades"
    post:
      tags:
      - "Actividades"
      summary: "Add one Actividad."
      description: "Add one Actividad."
      operationId: "addActividad"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - in: "body"
        name: "Actividad"
        required: true
        schema:
          $ref: "#/definitions/Actividades"
      responses:
        200:
          description: "Request OK."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Vacas"
        204:
          description: "Response is empty."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
        400:
          description: "Some parameters are missing or badly entered."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "string"
              format: "date-time"
              description: "The UTC date/time at which the current rate limit window\
                \ / resets."
          schema:
            type: "string"
        404:
          description: "Entity not found."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Illegal input for operation."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
      security:
      - appToken: []
      - userToken: []
      x-swagger-router-controller: "Actividades"
    put:
      tags:
      - "Actividades"
      summary: "Edit one Actividad."
      description: "Edit one Actividad"
      operationId: "editActividad"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - in: "body"
        name: "Lote"
        required: true
        schema:
          $ref: "#/definitions/Vacas"
      responses:
        200:
          description: "Request OK."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Actividades"
        204:
          description: "Response is empty."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
        400:
          description: "Some parameters are missing or badly entered."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "string"
              format: "date-time"
              description: "The UTC date/time at which the current rate limit window\
                \  resets."
          schema:
            type: "string"
        404:
          description: "Entity not found."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Illegal input for operation."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
      security:
      - appToken: []
      - userToken: []
      x-swagger-router-controller: "Actividades"
  /Actividades/{id}:
    get:
      tags:
      - "Actividades"
      summary: "Get one Actividad."
      description: "Get one Actividades."
      operationId: "getActividadById"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "id"
        in: "path"
        description: "id to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Request OK."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Actividades"
        204:
          description: "Response is empty."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
        400:
          description: "Some parameters are missing or badly entered."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "string"
              format: "date-time"
              description: "The UTC date/time at which the current rate limit window\
                \  resets."
          schema:
            type: "string"
        404:
          description: "Entity not found."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Illegal input for operation."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
      security:
      - appToken: []
      - userToken: []
      x-swagger-router-controller: "Actividades"
    delete:
      tags:
      - "Actividades"
      summary: "Delete one Actividad."
      description: "Delete one Actividades."
      operationId: "deleteActividad"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "id"
        in: "path"
        description: "id to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Request OK."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/DeletedResponse"
        204:
          description: "Response is empty."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
        400:
          description: "Some parameters are missing or badly entered."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "string"
              format: "date-time"
              description: "The UTC date/time at which the current rate limit window\
                \  resets."
          schema:
            type: "string"
        404:
          description: "Entity not found."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Illegal input for operation."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
      security:
      - appToken: []
      - userToken: []
      x-swagger-router-controller: "Actividades"
  /Actividades/upload:
    post:
      tags:
      - "Actividades"
      summary: "Add actividades from a file."
      description: "Add actividades from a file."
      operationId: "addActividades"
      consumes:
      - "multipart/form-data"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "upfile"
        in: "formData"
        description: "The file to upload."
        required: false
        type: "file"
      - name: "userId"
        in: "query"
        description: "id of user. Only for admin users."
        required: false
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Request OK."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/UploadResponse"
        204:
          description: "Response is empty."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
        400:
          description: "Some parameters are missing or badly entered."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "string"
              format: "date-time"
              description: "The UTC date/time at which the current rate limit window\
                \ / resets."
          schema:
            type: "string"
        404:
          description: "Entity not found."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Illegal input for operation."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
      security:
      - appToken: []
      - userToken: []
      x-swagger-router-controller: "Actividades"
...
securityDefinitions:
  appToken:
    type: "apiKey"
    name: "x_app_id"
    in: "header"
  userToken:
    type: "apiKey"
    name: "x_user_key"
    in: "header"
definitions:
  Lotes:
    type: "object"
    properties:
      _id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      latitude:
        type: "number"
        format: "float"
      longitude:
        type: "number"
        format: "float"
    example:
      latitude: 6.0274563
      name: "Campo Baires"
      id: 0
      longitude: 1.4658129
  Vacas:
    type: "object"
    properties:
      _id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      sex:
        type: "string"
      months:
        type: "integer"
        format: "int64"
      weight:
        type: "number"
        format: "float"
      location:
        $ref: "#/definitions/Lotes"
      actividades:
        type: "array"
        items:
          $ref: "#/definitions/Actividades"
    example:
      name: "Mirta"
      weigth: 350
      months: 30
      location:
        latitude: 6.0274563
        name: "name"
        id: 0
        longitude: 1.4658129
      id: 0
  Actividades:
    type: "object"
    properties:
      _id:
        type: "integer"
        format: "int64"
      sampleDate:
        type: "string"
        format: "date-time"
      latitude:
        type: "number"
        format: "float"
      longitude:
        type: "number"
        format: "float"
      vaca:
        $ref: "#/definitions/Vacas"
      insertDate:
        type: "string"
        format: "date-time"
      colectionDate:
        type: "string"
        format: "date-time"
      hdop:
        type: "number"
        format: "float"
      satellites:
        type: "integer"
        format: "int64"
      signal:
        type: "number"
        format: "float"
      voltage:
        type: "number"
        format: "float"
      tag:
        type: "string"
      connectionsAttempts:
        type: "integer"
        format: "int64"
...
  UploadResponse:
    type: "object"
    properties:
      actividades_insertadas:
        type: "integer"
        format: "int64"
      nuevas_vacas:
        type: "array"
        items:
          $ref: "#/definitions/Vacas"
      lineas_con_error:
        type: "array"
        items:
          $ref: "#/definitions/Linea"
    example:
      nuevas_vacas:
      - weigth: 1.4658129
        months: 6.02745618307040320615897144307382404804229736328125
        name: "name"
        location:
          latitude: 6.0274563
          name: "name"
          id: 0
          longitude: 1.4658129
        _id: 0
        actividades:
        - date: "2000-01-23T04:56:07.000+00:00"
          latitude: 5.962134
          longitude: 5.637377
        - date: "2000-01-23T04:56:07.000+00:00"
          latitude: 5.962134
          longitude: 5.637377
      - weigth: 1.4658129
        months: 6.02745618307040320615897144307382404804229736328125
        name: "name"
        location:
          latitude: 6.0274563
          name: "name"
          id: 0
          longitude: 1.4658129
        _id: 0
        actividades:
        - date: "2000-01-23T04:56:07.000+00:00"
          latitude: 5.962134
          longitude: 5.637377
        - date: "2000-01-23T04:56:07.000+00:00"
          latitude: 5.962134
          longitude: 5.637377
      lineas_con_error:
      - linea_number: 6
        linea: "linea"
      - linea_number: 6
        linea: "linea"
      actividades_insertadas: 0
  Linea:
    type: "object"
    properties:
      linea_number:
        type: "integer"
        format: "int64"
      linea:
        type: "string"
    example:
      linea_number: 6
      linea: "linea"
  Error:
    type: "object"
    required:
    - "message"
    properties:
      code:
        type: "string"
      message:
        type: "string"
      data:
        type: "string"
    example:
      code: "code"
      data: "data"
      message: "message"
  DeletedResponse:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
    example:
      id: 0